Copy release notes only if provided

+++ scripts/copy-cf-mgmt-release.sh
@@ -8,4 +8,6 @@ VERSION=$(cat ${SOURCE_FOLDER}/version)
 
 cp ${SOURCE_FOLDER}/cf-mgmt-config-linux ${OUTPUT_FOLDER}/cf-mgmt-config-linux-${VERSION}
 cp ${SOURCE_FOLDER}/cf-mgmt-linux ${OUTPUT_FOLDER}/cf-mgmt-linux-${VERSION}
-cp ${SOURCE_FOLDER}/body ${OUTPUT_FOLDER}/cf-mgmt-${VERSION}-release.md
+if [ -f ${SOURCE_FOLDER}/body ]; then
+  cp ${SOURCE_FOLDER}/body ${OUTPUT_FOLDER}/cf-mgmt-${VERSION}-release.md
+fi
