+++ scripts/fetch-cf-cli-release.sh
@@ -8,4 +8,6 @@ VERSION=$(cat ${SOURCE_FOLDER}/version)
 
 curl -Lo ${OUTPUT_FOLDER}/cf-cli-installer_${VERSION}_x86-64.rpm "https://packages.cloudfoundry.org/stable?release=redhat64&version=${VERSION}&source=github-rel"
 curl -Lo ${OUTPUT_FOLDER}/cf-cli-installer_${VERSION}_winx64.zip "https://packages.cloudfoundry.org/stable?release=windows64&version=${VERSION}&source=github-rel"
-cp ${SOURCE_FOLDER}/body ${OUTPUT_FOLDER}/cf-cli-${VERSION}-release.md
+if [ -f ${SOURCE_FOLDER}/body ]; then
+  cp ${SOURCE_FOLDER}/body ${OUTPUT_FOLDER}/cf-cli-${VERSION}-release.md
+fi
