+++ scripts/fetch-mc-release.sh
@@ -7,4 +7,6 @@ OUTPUT_FOLDER=$2
 VERSION=$(cat ${SOURCE_FOLDER}/version)
 
 curl -Lo ${OUTPUT_FOLDER}/mc-${VERSION} "https://dl.minio.io/client/mc/release/linux-amd64/mc"
-cp ${SOURCE_FOLDER}/body ${OUTPUT_FOLDER}/mc-${VERSION}-release.md
+if [ -f ${SOURCE_FOLDER}/body ]; then
+  cp ${SOURCE_FOLDER}/body ${OUTPUT_FOLDER}/mc-${VERSION}-release.md
+fi
